Fixing type bug

+++ src/shader.hpp
@@ -121,7 +121,7 @@ class Shader
 
     void setFloat(const std::string &name, float value) const
     {
-        glUniform1f(glGetUniformLocation(ID, name.c_str()), (int)value);
+        glUniform1f(glGetUniformLocation(ID, name.c_str()), (float)value);
     }
 
     void setVec2(const std::string &name, const glm::vec2 &value) const
